Trait sui_json_rpc::api::EventReadApiServer

pub trait EventReadApiServer: Sized + Send + Sync + 'static {
    fn get_events<'life0, 'async_trait>(
        &'life0 self,
        query: EventQuery,
        cursor: Option<EventID>,
        limit: Option<usize>,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<EventPage>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the EventReadApi RPC API.

Required Methods

fn get_events<'life0, 'async_trait>(
&'life0 self,
query: EventQuery,
cursor: Option<EventID>,
limit: Option<usize>,
descending_order: Option<bool>
) -> Pin<Box<dyn Future<Output = RpcResult<EventPage>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return list of events for a specified query criteria.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.