Trait sui_json_rpc::api::ReadApiClient

pub trait ReadApiClient: ClientT {

    // Provided methods
    fn get_owned_objects<'life0, 'async_trait>(
        &'life0 self,
        address: SuiAddress,
        query: Option<SuiObjectResponseQuery>,
        cursor: Option<ObjectID>,
        limit: Option<usize>,
        at_checkpoint: Option<CheckpointId>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ObjectsPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_dynamic_fields<'life0, 'async_trait>(
        &'life0 self,
        parent_object_id: ObjectID,
        cursor: Option<ObjectID>,
        limit: Option<usize>
    ) -> Pin<Box<dyn Future<Output = RpcResult<DynamicFieldPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_total_transaction_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transactions_in_range_deprecated<'life0, 'async_trait>(
        &'life0 self,
        start: TxSequenceNumber,
        end: TxSequenceNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionDigest>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_transaction_with_options<'life0, 'async_trait>(
        &'life0 self,
        digest: TransactionDigest,
        options: Option<SuiTransactionResponseOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_object_with_options<'life0, 'async_trait>(
        &'life0 self,
        object_id: ObjectID,
        options: Option<SuiObjectDataOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiObjectResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn multi_get_object_with_options<'life0, 'async_trait>(
        &'life0 self,
        object_ids: Vec<ObjectID>,
        options: Option<SuiObjectDataOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectResponse>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_dynamic_field_object<'life0, 'async_trait>(
        &'life0 self,
        parent_object_id: ObjectID,
        name: DynamicFieldName
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiObjectResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_move_function_arg_types<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module: String,
        function: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_normalized_move_modules_by_package<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_normalized_move_module<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_normalized_move_struct<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        struct_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_normalized_move_function<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        function_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn query_transactions<'life0, 'async_trait>(
        &'life0 self,
        query: SuiTransactionResponseQuery,
        cursor: Option<TransactionDigest>,
        limit: Option<usize>,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn multi_get_transactions_with_options<'life0, 'async_trait>(
        &'life0 self,
        digests: Vec<TransactionDigest>,
        options: Option<SuiTransactionResponseOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiTransactionResponse>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn try_get_past_object<'life0, 'async_trait>(
        &'life0 self,
        object_id: ObjectID,
        version: SequenceNumber,
        options: Option<SuiObjectDataOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiPastObjectResponse>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn try_multi_get_past_objects<'life0, 'async_trait>(
        &'life0 self,
        past_objects: Vec<SuiGetPastObjectRequest>,
        options: Option<SuiObjectDataOptions>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiPastObjectResponse>>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_latest_checkpoint_sequence_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<CheckpointSequenceNumber>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_checkpoint<'life0, 'async_trait>(
        &'life0 self,
        id: CheckpointId
    ) -> Pin<Box<dyn Future<Output = RpcResult<Checkpoint>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_checkpoints<'life0, 'async_trait>(
        &'life0 self,
        cursor: Option<CheckpointSequenceNumber>,
        limit: Option<usize>,
        descending_order: bool
    ) -> Pin<Box<dyn Future<Output = RpcResult<CheckpointPage>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }

}

Client implementation for the ReadApi RPC API.

Provided Methods

fn get_owned_objects<'life0, 'async_trait>( &'life0 self, address: SuiAddress, query: Option<SuiObjectResponseQuery>, cursor: Option<ObjectID>, limit: Option<usize>, at_checkpoint: Option<CheckpointId> ) -> Pin<Box<dyn Future<Output = RpcResult<ObjectsPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the list of objects owned by an address.

fn get_dynamic_fields<'life0, 'async_trait>( &'life0 self, parent_object_id: ObjectID, cursor: Option<ObjectID>, limit: Option<usize> ) -> Pin<Box<dyn Future<Output = RpcResult<DynamicFieldPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the list of dynamic field objects owned by an object.

fn get_total_transaction_number<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the total number of transactions known to the server.

fn get_transactions_in_range_deprecated<'life0, 'async_trait>( &'life0 self, start: TxSequenceNumber, end: TxSequenceNumber ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionDigest>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return list of transaction digests within the queried range. This method will be removed before April 2023, please use queryTransactions instead

fn get_transaction_with_options<'life0, 'async_trait>( &'life0 self, digest: TransactionDigest, options: Option<SuiTransactionResponseOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the transaction response object.

fn get_object_with_options<'life0, 'async_trait>( &'life0 self, object_id: ObjectID, options: Option<SuiObjectDataOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<SuiObjectResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the object information for a specified object

fn multi_get_object_with_options<'life0, 'async_trait>( &'life0 self, object_ids: Vec<ObjectID>, options: Option<SuiObjectDataOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectResponse>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the object data for a list of objects

fn get_dynamic_field_object<'life0, 'async_trait>( &'life0 self, parent_object_id: ObjectID, name: DynamicFieldName ) -> Pin<Box<dyn Future<Output = RpcResult<SuiObjectResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the dynamic field object information for a specified object

fn get_move_function_arg_types<'life0, 'async_trait>( &'life0 self, package: ObjectID, module: String, function: String ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the argument types of a Move function, based on normalized Type.

fn get_normalized_move_modules_by_package<'life0, 'async_trait>( &'life0 self, package: ObjectID ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return structured representations of all modules in the given package

fn get_normalized_move_module<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move module

fn get_normalized_move_struct<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String, struct_name: String ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move struct

fn get_normalized_move_function<'life0, 'async_trait>( &'life0 self, package: ObjectID, module_name: String, function_name: String ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a structured representation of Move function

fn query_transactions<'life0, 'async_trait>( &'life0 self, query: SuiTransactionResponseQuery, cursor: Option<TransactionDigest>, limit: Option<usize>, descending_order: Option<bool> ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return list of transactions for a specified query criteria.

fn multi_get_transactions_with_options<'life0, 'async_trait>( &'life0 self, digests: Vec<TransactionDigest>, options: Option<SuiTransactionResponseOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiTransactionResponse>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns an ordered list of transaction responses The method will throw an error if the input contains any duplicate or the input size exceeds QUERY_MAX_RESULT_LIMIT

fn try_get_past_object<'life0, 'async_trait>( &'life0 self, object_id: ObjectID, version: SequenceNumber, options: Option<SuiObjectDataOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<SuiPastObjectResponse>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Note there is no software-level guarantee/SLA that objects with past versions can be retrieved by this API, even if the object and version exists/existed. The result may vary across nodes depending on their pruning policies. Return the object information for a specified version

fn try_multi_get_past_objects<'life0, 'async_trait>( &'life0 self, past_objects: Vec<SuiGetPastObjectRequest>, options: Option<SuiObjectDataOptions> ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiPastObjectResponse>>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Note there is no software-level guarantee/SLA that objects with past versions can be retrieved by this API, even if the object and version exists/existed. The result may vary across nodes depending on their pruning policies. Return the object information for a specified version

fn get_latest_checkpoint_sequence_number<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = RpcResult<CheckpointSequenceNumber>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return the sequence number of the latest checkpoint that has been executed

fn get_checkpoint<'life0, 'async_trait>( &'life0 self, id: CheckpointId ) -> Pin<Box<dyn Future<Output = RpcResult<Checkpoint>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return a checkpoint

fn get_checkpoints<'life0, 'async_trait>( &'life0 self, cursor: Option<CheckpointSequenceNumber>, limit: Option<usize>, descending_order: bool ) -> Pin<Box<dyn Future<Output = RpcResult<CheckpointPage>> + Send + 'async_trait>>where Self: Sync + 'async_trait, 'life0: 'async_trait,

Return paginated list of checkpoints