Trait sui_json_rpc::api::RpcFullNodeReadApiServer

pub trait RpcFullNodeReadApiServer: Sized + Send + Sync + 'static {

    fn dev_inspect_transaction<'life0, 'async_trait>(
        &'life0 self,
        sender_address: SuiAddress,
        tx_bytes: Base64,
        gas_price: Option<u64>,
        epoch: Option<EpochId>
    ) -> Pin<Box<dyn Future<Output = RpcResult<DevInspectResults>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn dry_run_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx_bytes: Base64
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionEffects>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_move_function_arg_types<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module: String,
        function: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_normalized_move_modules_by_package<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_normalized_move_module<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_normalized_move_struct<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        struct_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_normalized_move_function<'life0, 'async_trait>(
        &'life0 self,
        package: ObjectID,
        module_name: String,
        function_name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_transactions<'life0, 'async_trait>(
        &'life0 self,
        query: TransactionQuery,
        cursor: Option<TransactionDigest>,
        limit: Option<usize>,
        descending_order: Option<bool>
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsPage>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn try_get_past_object<'life0, 'async_trait>(
        &'life0 self,
        object_id: ObjectID,
        version: SequenceNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<GetPastObjectDataResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_latest_checkpoint_sequence_number(
        &self
    ) -> RpcResult<CheckpointSequenceNumber>;
    fn get_checkpoint_summary(
        &self,
        sequence_number: CheckpointSequenceNumber
    ) -> RpcResult<CheckpointSummary>;
    fn get_checkpoint_summary_by_digest(
        &self,
        digest: CheckpointDigest
    ) -> RpcResult<CheckpointSummary>;
    fn get_checkpoint_contents(
        &self,
        sequence_number: CheckpointSequenceNumber
    ) -> RpcResult<CheckpointContents>;
    fn get_checkpoint_contents_by_digest(
        &self,
        digest: CheckpointContentsDigest
    ) -> RpcResult<CheckpointContents>;

    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the RpcFullNodeReadApi RPC API.

Required Methods

fn dev_inspect_transaction<'life0, 'async_trait>(
&'life0 self,
sender_address: SuiAddress,
tx_bytes: Base64,
gas_price: Option<u64>,
epoch: Option<EpochId>
) -> Pin<Box<dyn Future<Output = RpcResult<DevInspectResults>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Runs the transaction in dev-inpsect mode. Which allows for nearly any transaction (or Move call) with any arguments. Detailed results are provided, including both the transaction effects and any return values.

fn dry_run_transaction<'life0, 'async_trait>(
&'life0 self,
tx_bytes: Base64
) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionEffects>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return transaction execution effects including the gas cost summary, while the effects are not committed to the chain.

fn get_move_function_arg_types<'life0, 'async_trait>(
&'life0 self,
package: ObjectID,
module: String,
function: String
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<MoveFunctionArgType>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the argument types of a Move function, based on normalized Type.

fn get_normalized_move_modules_by_package<'life0, 'async_trait>(
&'life0 self,
package: ObjectID
) -> Pin<Box<dyn Future<Output = RpcResult<BTreeMap<String, SuiMoveNormalizedModule>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return structured representations of all modules in the given package

fn get_normalized_move_module<'life0, 'async_trait>(
&'life0 self,
package: ObjectID,
module_name: String
) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedModule>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return a structured representation of Move module

fn get_normalized_move_struct<'life0, 'async_trait>(
&'life0 self,
package: ObjectID,
module_name: String,
struct_name: String
) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedStruct>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return a structured representation of Move struct

fn get_normalized_move_function<'life0, 'async_trait>(
&'life0 self,
package: ObjectID,
module_name: String,
function_name: String
) -> Pin<Box<dyn Future<Output = RpcResult<SuiMoveNormalizedFunction>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return a structured representation of Move function

fn get_transactions<'life0, 'async_trait>(
&'life0 self,
query: TransactionQuery,
cursor: Option<TransactionDigest>,
limit: Option<usize>,
descending_order: Option<bool>
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionsPage>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return list of transactions for a specified query criteria.

fn try_get_past_object<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID,
version: SequenceNumber
) -> Pin<Box<dyn Future<Output = RpcResult<GetPastObjectDataResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Note there is no software-level guarantee/SLA that objects with past versions can be retrieved by this API, even if the object and version exists/existed. The result may vary across nodes depending on their pruning policies. Return the object information for a specified version

fn get_latest_checkpoint_sequence_number(
&self
) -> RpcResult<CheckpointSequenceNumber>

Return the sequence number of the latest checkpoint that has been executed

fn get_checkpoint_summary(
&self,
sequence_number: CheckpointSequenceNumber
) -> RpcResult<CheckpointSummary>

Return a checkpoint summary based on a checkpoint sequence number

fn get_checkpoint_summary_by_digest(
&self,
digest: CheckpointDigest
) -> RpcResult<CheckpointSummary>

Return a checkpoint summary based on checkpoint digest

fn get_checkpoint_contents(
&self,
sequence_number: CheckpointSequenceNumber
) -> RpcResult<CheckpointContents>

Return contents of a checkpoint, namely a list of execution digests

fn get_checkpoint_contents_by_digest(
&self,
digest: CheckpointContentsDigest
) -> RpcResult<CheckpointContents>

Return contents of a checkpoint based on checkpoint content digest

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.