Trait sui_json_rpc::api::RpcReadApiServer

pub trait RpcReadApiServer: Sized + Send + Sync + 'static {
    fn get_objects_owned_by_address<'life0, 'async_trait>(
        &'life0 self,
        address: SuiAddress
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectInfo>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_objects_owned_by_object<'life0, 'async_trait>(
        &'life0 self,
        object_id: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectInfo>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_dynamic_fields<'life0, 'async_trait>(
        &'life0 self,
        parent_object_id: ObjectID,
        cursor: Option<ObjectID>,
        limit: Option<usize>
    ) -> Pin<Box<dyn Future<Output = RpcResult<DynamicFieldPage>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_total_transaction_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_transactions_in_range<'life0, 'async_trait>(
        &'life0 self,
        start: TxSequenceNumber,
        end: TxSequenceNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionDigest>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_transaction<'life0, 'async_trait>(
        &'life0 self,
        digest: TransactionDigest
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_transaction_auth_signers<'life0, 'async_trait>(
        &'life0 self,
        digest: TransactionDigest
    ) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionAuthSignersResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_object<'life0, 'async_trait>(
        &'life0 self,
        object_id: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectDataResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_dynamic_field_object<'life0, 'async_trait>(
        &'life0 self,
        parent_object_id: ObjectID,
        name: String
    ) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectDataResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the RpcReadApi RPC API.

Required Methods

fn get_objects_owned_by_address<'life0, 'async_trait>(
&'life0 self,
address: SuiAddress
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectInfo>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the list of objects owned by an address.

fn get_objects_owned_by_object<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<SuiObjectInfo>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the list of objects owned by an object.

fn get_dynamic_fields<'life0, 'async_trait>(
&'life0 self,
parent_object_id: ObjectID,
cursor: Option<ObjectID>,
limit: Option<usize>
) -> Pin<Box<dyn Future<Output = RpcResult<DynamicFieldPage>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the list of dynamic field objects owned by an object.

fn get_total_transaction_number<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the total number of transactions known to the server.

fn get_transactions_in_range<'life0, 'async_trait>(
&'life0 self,
start: TxSequenceNumber,
end: TxSequenceNumber
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<TransactionDigest>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return list of transaction digests within the queried range.

fn get_transaction<'life0, 'async_trait>(
&'life0 self,
digest: TransactionDigest
) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the transaction response object.

fn get_transaction_auth_signers<'life0, 'async_trait>(
&'life0 self,
digest: TransactionDigest
) -> Pin<Box<dyn Future<Output = RpcResult<SuiTransactionAuthSignersResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the authority public keys that commits to the authority signature of the transaction.

fn get_object<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID
) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectDataResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the object information for a specified object

fn get_dynamic_field_object<'life0, 'async_trait>(
&'life0 self,
parent_object_id: ObjectID,
name: String
) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectDataResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Return the dynamic field object information for a specified object

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.