Module sui_json_rpc::api

Structs

CoinReadApiOpenRpc

EventReadApiOpenRpc

EventStreamingApiOpenRpc

GovernanceReadApiOpenRpc

RpcBcsApiOpenRpc

RpcFullNodeReadApiOpenRpc

RpcReadApiOpenRpc

RpcTransactionBuilderOpenRpc

ThresholdBlsApiOpenRpc

TransactionExecutionApiOpenRpc

TransactionStreamingApiOpenRpc

Constants

QUERY_MAX_RESULT_LIMIT

Maximum number of events returned in an event query. This is equivalent to EVENT_QUERY_MAX_LIMIT in sui-storage crate. To avoid unnecessary dependency on that crate, we have a reference here for document purposes.

Traits

CoinReadApiClient

Client implementation for the CoinReadApi RPC API.

CoinReadApiServer

Server trait implementation for the CoinReadApi RPC API.

EventReadApiClient

Client implementation for the EventReadApi RPC API.

EventReadApiServer

Server trait implementation for the EventReadApi RPC API.

EventStreamingApiClient

Client implementation for the EventStreamingApi RPC API.

EventStreamingApiServer

Server trait implementation for the EventStreamingApi RPC API.

GovernanceReadApiClient

Client implementation for the GovernanceReadApi RPC API.

GovernanceReadApiServer

Server trait implementation for the GovernanceReadApi RPC API.

RpcBcsApiClient

Client implementation for the RpcBcsApi RPC API.

RpcBcsApiServer

Server trait implementation for the RpcBcsApi RPC API.

RpcFullNodeReadApiClient

Client implementation for the RpcFullNodeReadApi RPC API.

RpcFullNodeReadApiServer

Server trait implementation for the RpcFullNodeReadApi RPC API.

RpcReadApiClient

Client implementation for the RpcReadApi RPC API.

RpcReadApiServer

Server trait implementation for the RpcReadApi RPC API.

RpcTransactionBuilderClient

Client implementation for the RpcTransactionBuilder RPC API.

RpcTransactionBuilderServer

Server trait implementation for the RpcTransactionBuilder RPC API.

ThresholdBlsApiClient

Client implementation for the ThresholdBlsApi RPC API.

ThresholdBlsApiServer

Server trait implementation for the ThresholdBlsApi RPC API.

TransactionExecutionApiClient

Client implementation for the TransactionExecutionApi RPC API.

TransactionExecutionApiServer

Server trait implementation for the TransactionExecutionApi RPC API.

TransactionStreamingApiClient

Client implementation for the TransactionStreamingApi RPC API.

TransactionStreamingApiServer

Server trait implementation for the TransactionStreamingApi RPC API.

Functions

cap_page_limit